+++ tests/compas/geometry/test_offset.py
@@ -1,9 +1,10 @@
 import pytest
 
+from compas.geometry import Line
+from compas.geometry import Polyline
+from compas.geometry import Polygon
+from compas.geometry.offset.offset import offset_polyline as c_offset_polyline
 from compas.geometry import allclose
-from compas.geometry import offset_line
-from compas.geometry import offset_polygon
-from compas.geometry import offset_polyline
 
 # ==============================================================================
 # polygon
@@ -22,8 +23,9 @@ from compas.geometry import offset_polyline
     ],
 )
 def test_offset_polygon(polygon, distance, tol, output_polygon):
+    input_polygon = Polygon(polygon)
     output_polygon = [v for v in output_polygon]
-    assert allclose(offset_polygon(polygon, distance, tol), output_polygon)
+    assert allclose(input_polygon.offset(distance, tol=tol), output_polygon)
 
 
 @pytest.mark.parametrize(
@@ -37,7 +39,7 @@ def test_offset_polygon(polygon, distance, tol, output_polygon):
                 [0.0, 1.0, 0.0],
                 [0.0, 0.5, 0.0],
             ],
-            [0.10],
+            0.10,
             1e-6,
             [
                 [0.1, 0.1, 0.0],
@@ -50,7 +52,8 @@ def test_offset_polygon(polygon, distance, tol, output_polygon):
     ],
 )
 def test_offset_colinear_polygon(polygon, distance, tol, output_polygon):
-    assert allclose(offset_polygon(polygon, distance, tol), output_polygon)
+    input_polygon = Polygon(polygon)
+    assert allclose(input_polygon.offset(distance, tol=tol), output_polygon)
 
 
 # ==============================================================================
@@ -63,8 +66,9 @@ def test_offset_colinear_polygon(polygon, distance, tol, output_polygon):
     [([[0.0, 0.0, 0.0], [1.0, 0.0, 0.0]], 1, [0.0, 0.0, 1.0], 1e-6)],
 )
 def test_offset_polyline_equals_offset_line(polyline, distance, normal, tol):
-    output_line = [v for v in offset_line(polyline, distance, normal)]
-    assert allclose(offset_polyline(polyline, distance, normal, tol), output_line)
+    input_polyline = Polyline(polyline)
+    output_polyline = [v for v in input_polyline.offset(distance, normal=normal)]
+    assert allclose(input_polyline.offset(distance, normal=normal, tol=tol), output_polyline)
 
 
 @pytest.mark.parametrize(
@@ -80,8 +84,9 @@ def test_offset_polyline_equals_offset_line(polyline, distance, normal, tol):
     ],
 )
 def test_variable_offset_on_colinear_polyline(polyline, distance, normal, tol, output_polyline):
+    input_polyline = Polyline(polyline)
     output_polyline = [v for v in output_polyline]
-    assert allclose(offset_polyline(polyline, distance, normal, tol), output_polyline)
+    assert allclose(c_offset_polyline(input_polyline, distance, normal=normal, tol=tol), output_polyline)
 
 
 # ==============================================================================
@@ -94,5 +99,6 @@ def test_variable_offset_on_colinear_polyline(polyline, distance, normal, tol, o
     [([[1.0, 0.0, 0.0], [1.0, 0.0, 0.0]], 1, [0.0, 0.0, 1.0])],
 )
 def test_offset_line_zero_length(line, distance, normal):
-    output_line = offset_line(line, distance, normal)
-    assert allclose(line, output_line)
+    input_line = Line(*line)
+    output_line = input_line.offset(distance, normal=normal)
+    assert allclose(input_line, output_line)